dver: Always bind-mount a /etc/dmd.conf
Ancient versions of the compiler read the configuration file only from
that location, so (besides DFLAGS, or explicit compiler arguments, or
patching the binary, or recompiling from source) that's the only way
to specify Phobos's location for these compiler versions.

Previously, we only bind-mounted /etc/dmd.conf if it existed on the
host, to prevent the system DMD version from messing with the version
we want to run. However, we need to do this regardless if the file
exists to support these old versions; so, just do it always.